Added user authorization on front-end

+++ backend/routes/posts.js
@@ -33,7 +33,8 @@ router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()
     .then(createdPost => {
@@ -82,29 +83,44 @@ router.get('/:id', (req, res, next) => {
   });
 });
 
-router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
-  Post.findOne({_id: req.params.id})
-  .then(document => {
+router.put(
+  '/:id',
+  checkAuth,
+  multer({storage: storage}).single("image"),
+  (req, res, next) => {
     let imagePath = req.body.imagePath;
+
     if (req.file) {
       const url = req.protocol + '://' + req.get('host');
-      imagePath = url + '/images/' + req.file.filename
+      imagePath = url + '/images/' + req.file.filename;
     }
 
-    document.title = req.body.title;
-    document.content = req.body.content;
-    document.imagePath = imagePath;
-
-    document.save().then(result =>{
-      res.status(200).json({ message: "Update Successful!", post: result });
+    const post = new Post({
+      _id: req.body.id,
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-  });
-});
+
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Update sucessful!' });
+      } else {
+        res.status(401).json({ message: 'Not Authorized' })
+      }
+    });
+  }
+);
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
-  .then(() => {
-    res.status(200).json({message: 'Post Deleted'});
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
+  .then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Update sucessful!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized' })
+    }
   });
 });
